AUTH-1273: create separate docker image for fargate
Rename existing local image and update docker-compose.yml to use it.
Add a new dockerfile to run the application in fargate. This image needs all files locally as it cannot rely on a mounted volume.
Use 'cp' instead of 'copyfiles' in one place in package.json as copyfiles was hanging when building and/or running the docker image.

A separate image is needed to run in fargate as we are currently using a CF node buildpack rather than a docker image. The existing local image could not be run in fargate as-is.