Remove cookies and feedback link
We are removing the stored cookie and feedback consent attributes from
GOV.UK's Account API and the management page in the parts of the account
home area that are served from GOV.UK's Frontend app [[1]].

Before we can remove that page, we need to remove the link to it here.

This is the only link to this page, so we can also remove the config
and environment variables that hold it.

[1]: https://github.com/alphagov/frontend#3225