GUA 473: Add new header to the account
Add new header design including the new brand name as well as links to
the services page (which does not exist yet).
The new design is behind a flag for now seeing as we might want to
switch everything on (the account area redesign and header redesign) all
at the same time.