BAU: Tag dev image with commit SHA
Previously this action would tag the Docker image as 'latest'. ECR
handles this ok by saving the newly uploaded image as 'latest' and
untagging the image that was previously 'latest'.

However when the CloudFormation template is applied, the running image
in ECS isn't updated. I think this is for one of two reasons:

1. CloudFormation sees that ECS is already running an image tagged
   'latest' and decides not to update it
2. ECS caches images and when the new task definition is started it uses
   the already cached 'latest' image, which isn't what we want.

Both of these can be prevented by tagging each image with a unique ID.