GUA-503: Show the user's new email when sending a new code
The only way a user can get a verification code sent to their email
address by this app is while they're going through the journey to change
their email address. The other credential and delete journeys require
them to re-enter their password, but we never send a code to their
email.

That means the only way a user can end up resending an email code is if
they're trying to send a code to their new email address while they're
in the process of switching. That means we should display their new
email address on this page.

Note - this doesn't change any behaviour about which email we send the
code to. This is already correct - it already correctly goes to their
new email address.

This change only switches the email displayed to the user to the right
one.