GUA-514: Save public subject ID to session
This is a new field returned in the userinfo response for this
application. The public subject ID is the user ID the app currently gets
as the subjectID, and it's also the user ID the GOV.UK relying party
gets.

We want to change the subject ID for this app to the DI internal sector
pairwise ID, but we also need the public subject ID so we can send
notifications to GOV.UK when someone updates or deletes their account.

The first step for this is to separate out the public subject ID into
it's own variable.