GUA-470: Refactor aws config functions
We'll need to make something very similar to this for SNS.

Let's avoid any potential confusion or overlap by giving this a more specific
name.

Also the SNS config is likely to share much of the AwsConfig payload, rather
than creating a nearly identical setup let's extract what we'll need from here
and allow it to be shared.