GUA-470: Add Generic SNS publication service
Trying to make this follow the established patterns as much as possible.  See
the setup for SSM and KMS.

I've not tested this explicitly with unit tests, this may be an oversight.
However the functioning of this section should be covered explicitly by higher
level tests on the service that will invoke this.

We may want to reassess that if we end up using this generic service to hit
other SNS topics too.