GUA-470: Add deleteServiceData to deleteAccountService
Following established patterns from the application we can add our service
store delete to the pre-existing element here that already wraps the call to
Auth / Orchestration's account deletion.

Down the road if we wanted to refactor the auth element off, I'd question if
this abstraction is useful to us, perhaps we could get the controller
interacting with the SNS topic more directly?

But for now let's follow the established patterns of the app.