GUA-462: Add dynamoDBService
Follows a pattern established previously in the app and built upon with
the SNS delete topic pattern.

Essentially give a generic interface that can handle any request,
wrapped for convenience. Then call it from a service with default params
etc.

Down the line we may want to check this pattern is giving us value for
the number of lines of code it represents. But for now going with this
for consistency.