BAU: fix header/nav issues
Firefox doesn't like the double negative margins between the DS header
and the new account nav which is leading to some bizzare behaviour when
zooming in and out where the nav links are jumping to the left/right at
different zoom levels on Firefox.

This implements an alternative solution which instead of hiding the blue
border at the bottom of the black bar by adding negative margins to the
account nav, uses a modifier class which removes the bar and negative
margins at the bottom of the header.

Also makes the container for the GOVUK logo and Account product name
take up 100% of the width of the containing element. It was spotted that
the "Account" product name collapses below the logo on certain
breakpoints. Increasing the width of the container ensures this will not
happen.