GUA-550: Remove NOTIFICATION_TYPE.VERIFY_EMAIL from resend-email-code-controller.ts
notificationType appears to no longer be needed as it's being imported directly
as part of the function.  We suspect that this was a change that was made
previously and not updated everywhere across the application.

As a result, we suspect the frontend application has been setting incorrect
values in some of the headers it sent to the change email endpoint and reset
email endpoint. We'll query with the Authentication Team about any potential
impact.