GUA-612: phone number journey content updates
Implement content updates on the Change phone number journey. These are
from designs from Auth which are on their Figma (journeys we inherited)

One of the things missing was an entire page (the resend phone code
journey). In addition to the content changes to bring us in line with
Auth's figma, this also adds the missing page to the journey.