GUA-687: Adjustments to SCSS imports
Optimise SCSS imports from the DS npm module so that only styles for
components being used in the application are imported. Previously we
were indiscriminately importing all styles including for components
which are not currently used in the account management app.

Add load path to sass so that DS scss imports automatically resolve to
their location in node_modules.