BAU: Remove ci-only github workflow
Move the validation checks into the main template that gets fired on
merge. We only ever want these ones taking `main` as the git ref so that
we're testing, building and deploying to build.

Dev Platform explicitly only wants you doing this against main.

This is part of work to handle bugs we're seeing where branches are not
being built and tested against dev the way we'd expect