BAU: Add On push branch CI
It's going to validate and perfom app checks, container checks and
cloud formation checks. The checks that require the remote env, will
always use our dev instance, and be separate from the actual secure
pipelines CI/CD that sends things to build.

This is important as the build env will never be happy with any request
that uses a git branch other than main.

We recognise that this is a bit ugly and there's lots of chances for a
general refactor, but we thought we'd start by adding duplication so
that it works, and then the opportunities for DRYing things out is
easily testable, and a separate small unit of work.