GUA-404: Delete all Auth Team Terraform
This app was one of the earliest started on the program. The Auth team
went with Terraform at the time for Infra as Code and CI. This worked
fine but the program later committed to CloudFormation and provisioning
through Secure Delivery Pipelines.

When we did the Switchover of Infrastructure in Winter 2022, we made
much of this redundant, it hung around to allow us to fallback deploy if
anything went wrong in our patch. However now their infra has gone too
so this can all be removed.