BAU: Check for pinned versions by SHA when running github actions
Why? https://julienrenaux.fr/2019/12/20/github-actions-security-risk/

But I agree with this:
https://michaelheap.com/improve-your-github-actions-security/

> I agree with Julien that using arbitary actions is a risk, but as
always it’s a compromise between security and making life easy for
ourselves. Specifying a commit hash each time we want to upgrade could
become painful very quickly, especially if you’re using a large number
of actions.

With that in mind, I thought about how we could solve the problem with
automation and came up with the following solution.

> TL;DR: Using GitHub actions with branch names or tags is unsafe.
Use commit hash instead

So let's add checks to pin github action versions to ones we trust
enough to use, and then run the pin-github-action[2] tool to pin the new
versions in this workflow

[1]: https://github.com/marketplace/actions/ensure-sha-pinned-actions
[2]: https://michaelheap.com/improve-your-github-actions-security/#using-pin-github-action