Bump P-memoize from v4.0.1 to v7.1.1
P-memoize is a caching startegy dependency, we use it in the utils/oidc
component [1], to prevent having lots of round trips to the .well-known
discovery endpoints involved in OIDC.

It is a production dependency, and failure to resolve those discovery
details would result in a failure to auth, so we should upgrade this
with more care.

Here's a summary of the version bumps:

[v4..v5](https://github.com/sindresorhus/p-memoize/releases/tag/v5.0.0)

Breaking

- Require Node.js 12.20 d95278e
- This package is now pure ESM. Please [read
  this](https://gist.github.com/sindresorhus/a39789f98801d908bbc7ff3ecc99d99c).
- pMemoize.clear() is now a named export called pMemoizeClear

The ESM approach will break our app we'll need to update that,

We don't seem to call pMemoize.clear() anywhere, so not relevant. And
we're already on v16.17.0 of node.

[v5..v6](https://github.com/sindresorhus/p-memoize/releases/tag/v6.0.0)

Breaking

- Remove maxAge option 19b418e
    - For a custom cache, this means instead of .set() being provided an object with the properties value and maxAge, it will only be provided value as the first argument.
      [How to preserve max age
      functionality.](https://github.com/sindresorhus/p-memoize#time-based-cache-expiration)

We use max age everywhere, will need to fix these.

[v6..v7](https://github.com/sindresorhus/p-memoize/releases/tag/v7.0.0)

Breaking

- Require Node.js 14 ac4f137
- Remove cachePromiseRejection option (#48) 7a82e0f
  - [Reason](https://github.com/sindresorhus/p-memoize/pull/48)
  - [How to cache promise rejections
    now](https://github.com/sindresorhus/p-memoize#caching-promise-rejections)

We're on node 16, and don't use cachePromiseRejection

[1]: https://github.com/alphagov/di-account-management-frontend/blob/387b7df378dcbae6e360f9c1d8b67389009fab69/src/utils/oidc.ts#L3