GUA-749: implement activity log page
Implement the frontend for the new activity log page.
Created a new view at /security/sign-in-history which currently works
with hard coded fixtures but will be populated with real data once the
backend is completed.
The new view renders a list of signed in events, with the first event
either being an account creation event, or an introductory paragraph if
the user is an existing user.