OLH-869: Show password component labels
Add distinct labels for the show password component on the second
password field on the "Enter your new password" page.

The first password component is still using the default labels while the
"confirm password" field is using labels to reflect that this is a
retyped password.

This is so that screen reader users can more clearly discern which of
the Show/Hide buttons is associated with each password field.