BAU: Update pull request template
Our team's Jira prefix is now OLH.

We don't need to put a link to the Jira ticket in the PR description -
as long as the branch, commit or PR title starts with the ticket ID (eg.
OLH-1234) then Jira will pick it up and link them automatically.

Remove the TODO link for application configuration.

Remove the requirement to tell the tech lead if the Dockerfile changes.
I don't think we've ever done that or need to.

Add a new testing section to describe any manual testing that's been
done.

Remove the other considerations section. If we think any of these are
necessary they should be part of the acceptance criteria on the Jira
ticket.

Add a new how to review section to explain any non-standard steps or
areas the author isn't sure about eg. asking for clarification on a
language feature or providing a sample Cloudformation stack.