BAU: Refactor tests which spy on the logger
The type system in Pino seems to have changed in v8 and Typescript is no
longer happy for us to pass a spy instance in the `req` object.

Refactor these tests so we spy on the logger separately, then pass the
logger through to `req`.

There might be a different way of doing this that involves setting up
sinks for the logger and spying on those instead, but that would mean
refactoring the whole logging system. We don't have great unit test
coverage for logging, so that felt both too risky and too much work for
a dependency update.