BAU: Switch to Alpine Linux as Docker base image
Our final Docker image we use in production is 1.27GB. This is causing
slow builds and deploys as it takes time to move the large image around.
We also pay AWS per GB in the image registry.

Switching to Alpine as the base rather than Debian reduces the final
image size to only 330MB. This will speed up deploys and save us money.