BAU: Update form-action directive in CSP
It was reported that when the /delete-account form was submitted, the
page would hang instead of redirecting to the confirmation page.

The POST request showed up in the network tab with the correct location
and status code, and the account deletion was performed correctly so we
narrowed down the issue as being a CSP one.

The issue came from a helmet upgrade where the `form-action` directive
is now set to `self` by default meaning that redirects to the oidc
subdomain were blocked.
https://github.com/helmetjs/helmet/blob/main/CHANGELOG.md#500---2022-01-02

This adds subdomains of account.gov.uk to the `form-action` directive
which should fix the issue.