BAU: Upgrade local Node versions
We upgraded the Docker images in #942 but that missed the local files.

I'm not sure if these are used, or if we need both of them, but it is
possible to run the app locally without Docker so we should match the
version of Node that we use when it's deployed.