BAU: Update action step name
We're no longer using Node 16 - see #990.

Update the step name in the action to remove the reference to the Node
version we're using. This means we won't need to update this again for
future upgrades.

Note this doesn't change which version of Node we're using to run any
tests; it's only the step name that appears in the action logs.