BAU: Handle new error types in express-validator
express-validator v7.0.0 has introduced new error types[[1]]. As part of
this they have also renamed the `param` property to `path`.

Use the new parameter name and update the formatter to use check which
error type we're dealing with.

If it's a `FieldValidationError` then that has the path property.

`AlternativeValidationError` and `GroupedAlternativeValidationError`
errors are nested or double-nested error types. In those cases, take the
path property from the first error in the list, as they should all be
the same.

[1]: https://express-validator.github.io/docs/api/validation-result#error-types