BAU: Update error types
The `response` property in the `AxiosError` type is now a generic type
instead of only a string. In this case `response.data` should always be
a string however, so it's safe to use `as string` later on to keep the
compiler happy.

That's because `AxiosError.response` is type `AxiosResponse<T = any, D = any>`
which is a key, value set of `any` to `any`.

And `AxiosResponse.data` is type `T`. Since `response.data` is the same
as calling `response['data']` on an object, our key here is a string
type so we should expect `response.data` to be the same type ie. a
string.