BAU: Enable the API Gateway 5XX error alarm
This alarm triggers when more than 1% of responses from API Gateway are
a 5XX status code, as long as there have been a minimum of 10 responses
in two out of the last 5 minutes.

The most likely reason for this alarm to trigger is the frontend either
being down completely or having some other problem that needs attention.

Right now this alarm sends a slack message to the
`#govuk-accounts-notifications` channel. In the future we should update
it to also send alarms in production to `#di-2nd-line`. This closes a
gap in our monitoring of the frontend to provide a proactive
notification to us if there's a problem.

We've had this alarm set up for the best part of a year with actions
enabled. I've checked the alarm history and I'm confident there haven't
been any false positives so it's safe to switch it on.