OLH-1113: Set up for triage page
This sets up the boilerplate for the new contact centre triage page
we'll be building. There's a route, controller and a simple integration
test.

The new page is hidden behind a feature flag and is enabled up to
staging so we can show our teammates what's going on and also set up the
audit event integration with TxMA there when we need to.