OLH-1115: Add reference code helper
This helper function generates the reference code that we'll display to
users and ask them to read out over the phone. It must be a six digit
number.

Separating this out into its own helper (rather than as part of the
controller) means we can be very clear about the format requirements in
the tests.