OLH-1115: Log the reference code and lookup data
When a users calls our contact centre they'll read out the reference
code to the agent. We need to also log this code along with information
about the user's browser and web session so we can link their GOV.UK One
Login journey to their contact centre ticket.

After we've collected all the data from params and the user's session,
emit a log event containing everything.

Note we are not logging the user's user ID (if they are logged in and
have one). This is because we treat user IDs as personal data and we're
not allowed to log them to Splunk / Cloudwatch.

Instead we will have to use their session IDs and match that to events
in Auth to get their internal journey ID, which can be used to see what
that user did inside GOV.Uk One login