BAU: Update triage page URL
We intended to put a hyphen in the URL here, and have told other teams
that the URL will be `...gov-uk...`. It's easier for us to update the
URL here rather than getting all the other teams to make another change.

This page isn't live yet so we don't need a redirect from the old to
new URL.