OLH-1168: Put the webchat widget on the triage page
Put the webchat widget on the triage page behind the existing feature
flag. The supplier's docs for the webchat ask us to put it in the
`<head>` section, but this is the same implementation as on the previous
webchat demo page - it works fine in the `<body>` and is much easier to
implement in our templates.