OLH-1174: Allow the frontend to send audit events
Import the audit event queue and key parameters added in [^1] from SSM.
Set the queue URL as an environment variable and modify the frontend's
IAM permissions so it can write to the queue. These permissions are the
same as our `format-*` lambdas have on the backend to write to their
output queues so they should cover what's needed.

The environment variable isn't used for anything yet - we'll generate
the audit events and send them in a future ticket.

[^1]: https://github.com/alphagov/di-account-management-backend/pull/152