OLH-1122: Emit an audit event when a user visits the triage page.
The values from the query params and session are now needed in more than one place in the controller so the first
thing we do is to update the session with any safe and valid values that may have been passed as query parameters.
Then the other functions in the controller can get all the data they need from the session in the request.

Changed the main function in the controller so that all it does is call local functions that do all the work.

The controller assumes that there will always be a session due to the express-session middleware being used therefore
there is no need to use optional chanining for the session in the request.

Added an sqsService function to utils.  The SqsService is called from the EventService send
fucntion.  The EventService does not specify the specific queue to send the Event to it just provides a JSON
format of the Event.  The SqsService will try to send any failed events to the DLQ.

Use the event to create the log message that a user has visited the page as it will always have the required
data.  Moved the build event function to be called before creating the log entry.  Renamed the function that
does the logging to reflect intention rather than how.  This refactor exposed an issue with the tests around
the user agent retrieval that is now always from the session after initial setup.