OLH-1165: Add an example client side JS test
There's a new npm script (`npm run test:js`) that runs mocha looking for
*.test.js files in the assets directory.

The example tests use the same `chai` and `mocha` runner and framework
that the Typescript tests use. We need to use CommonJS syntax rather
than the `import {expect} from "chai"` ES6 syntax, but that's a
difference we can manage.

I've not yet added this script to the list of ones which are run in `npm
run test` - we can do that once we have some more substantial tests.

Added an example client side JS test - simple test of a function in cookies.js