OLH-1122: Emit an audit event when a user visits the triage page - refactoring and extra unit test coverage.
The values from the query params and session are now needed in more than one place in the controller so the first
thing we do is to update the session with any safe and valid values that may have been passed as query parameters.
Then the other functions in the controller can get all the data they need from the session in the request.

Changed the main function so that all it does is call local functions that do all the work.

The controller assumes that there will always be a session due to the express-session middleware being used therefore
there is no need to use optional chanining for the session in the request.