OLH-1122: Emit an audit event when a user visits the triage page - adding SQS integration.
Initial commit adding an sqsService function to utils.  The SqsService is called from the EventService send
fucntion.  The EventService does not specify the specific queue to send the Event to it just provides a JSON
format of the Event.  The SqsService will try to send any failed events to the DLQ.