OLH-1122: Emit an audit event when a user visits the triage page - improving the logging.
Use the event to create the log message that a user has visited the page as it will always have the required
data.  Moved the build event function to be called before creating the log entry.  Renamed the function that
does the logging to reflect intention rather than how.  This refactor exposed an issue with the tests around
the user agent retrieval that is now always from the session after initial setup.