OLH-1183: Update content for triage page
Implement updates to the content of the new triage page. Add Welsh
content.
In order to support a phased approach to the rollout of this page, the
structure of the content has changed slightly. We now display slightly
different versions of the content depending on which feature flag(s) are
currently enabled.