OLH-1209: "Use webchat" conditional display
Display the "Use webchat" link only once Webchat is actually available
to use.

There are a number of things that could potentially go wrong: poor
network conditions meaning that the chat is slow to load, JS disabled in
the browser, or a breaking error with the chat preventing it from
loading for instance. In all of these instances, displaying the "Use
webchat" link unconditionally could result in users attempting to
interact with it and having a frustrating experience when clicking it
does not produce any outcome.

To display the link conditionally, I listened for the CHAT_LOADED event
dispatched by the webchat script. Seeing as the webchat script loads
somewhat asynchronously however, that means that the "SaEvents" object
is not immediately available to subscribe to. The way I got around this
is I set an interval checking for the existence of the object. If the
object still does not exist after 10 seconds of trying, break out of the
loop.

To hide the "Use webchat" link button, I used visibility: hidden as we
wanted the layout to remain unchanged to prevent a layout shift when the
element becomes visible.