OLH-1190: Don't export logs to Splunk in build
We run our performance (load) tests in build and these generate several
GB of logs each run. We share the daily Splunk ingest quota with all of
Cabinet Office and we're regularly hitting that quota which means we've
been asked to stop sending logs to Splunk when we run performance tests.

It's much simpler to simply not send logs to Splunk in build at all. If
we need to look at application logs for any reason we still have them in
Cloudwatch, or we can try to replicate an issue in a higher environment
where we do export the logs.