BAU: Enable webchat in integration
We thought our suppliers had access to the GDS VPN and therefore could
access the webchat in our IP-restricted staging environment. That's
turned out not to be true.

We could gather their IP ranges and add them to the allowlist, but it's
easier to simply enable the webchat in integration where we don't
restrict by IP address.