OLH-1261: Use the latest version of localstack
in the Github actions pipeline.

This is required because we use the `ubuntu-latest` action runner which
has just switched over to a new version of Ubuntu. The new version has a
boto that's incompatible with the old version of localstack we were
using previously.

Note this doesn't change the version of localstack we use for local
development. This version works in the pipeline, and the main priority
is getting that unblocked. We can come back and upgrade the version in
`docker-compose.yml` later.