OLH-1262: Include user IDs in triage page audit events
Add two new fields to the audit events emitted by the triage page -
`user.user_id` and `extensions.is_signed_in`.

The Technical Service Desk have requested we add the user ID to help
them match events between sessions / users. The user ID will be hashed
by TxMA, so we also add a new field to the unhashed `extensions` object
that will tell TSD if a user is signed in when we sent the event.

This also changes the string we send for session etc. IDs when a user is
signed out. This will make it more obvious to TSD if the session is
missing because of a bug when sending the event, or if the user is
signed out.