BAU: Rename application
We renamed this repository ages ago but missed the Node package name.
This hasn't mattered at all up until now, but we've just onboarded onto
Dynatrace and it's picking up the service name from this app's package
name.

This has the potential to be confusing when people from outside our team
are looking at Dynatrace and trying to figure out who to contact if
there's an incident.