OLH-1279 Investigate reference codes with multiple session ids.
Updating logging messages to include a trace id to link a users log messages easily.  Using session id for the value of the trace.
Tidied the app.constants ERROR_MESSAGES constant to remove unused messages.
Introduced new constant to app.constants for info log messages.
Added additional logging to logout-controller to see if the failure to destroy the session is causing the issue with the reference code persisting between sessions.