OLH-1312: update contact page email section
The opening hours for the email route have changed now that it is
managed by the supplier rather than our in house team.
This updates the information on the contact page to the most up to date
one. The hours changed on the 12th December so this update is overdue.