OLH-1304: Ensure chat icon scales better on zoom
When zoom is applied, the chat button takes up a very large proportion
of the screen. There is also a separate issue when text zoom is applied
where the text inside the notification counter becomes hard to read.
This is an attempt to fix both those issues and making the button scale
better by using relative units to set max and min dimensions on the
relevant elements.