OLH-1337 Redact user id from txma error log.
During development of this change I noticed that the tests in the test/unit folders were being excluded from the npm run test:unit script. Some of these tests were for modules changed for the PR and they were moved into the src folder structure to ensure they get run.
Where I found logging statements that did not use a message constant from app.constants.ts I added them.
Rather than using the export command as a prefix I standardised on using a final export statement for the modules changed in this PR as this seems to be standard across most of the code base and makes it easier to see what is meant to be public in the module.