OLH-1363: Custom nonce approach for webchat styles
The webchat supplier have updated the webchat so that it will work with
our content security policy.

We can now include the webchat stylesheets manually into our code with
the nonce generated by us. This implements that change, adding the
stylesheets to the <head> conditionally only when webchat is enabled,
and storing the paths as an environment variable.

The href values will be different per environment but the webchat
supplier have not supplied the values for other environments yet, so
it's only dev for now.
Once they have confirmed the values for other
environments, the environment variables can be updated.