OLH-1375: Add high traffic warning on triage page
We have been asked to publish "emergency messaging" on the contact
triage page in the event that the contact centre is overwhelmed by
calls.
This is a preemptive measure ahead of the Veterans ID card service, a
high profile service, onboarding with OL soon.
Welsh language version still TBD. This adds the English language version
of the content in the Design System "warning text" style.