OLH-1363: Refactor webchat URLs
Clean up the two script and CSS imports.
Both of these assets are served from the same domain and we'll always
want to use the same environment for each of them.
This changes the environment variable to be the base domain and then
builds the asset paths in the template.

This means we can get rid of an environment variable and the associated
boilerplate in `config.ts`