OLH-1505: Refactor date formatting function
The prettifyDate util was hard coded to return dates in en-GB. There is
a need to display dates in Welsh when the user is viewing their OL home
in Welsh.
This refactors the formatter to allow an additional locale parameter to
be passed in, to display dates in Welsh when the language has been set
to Welsh.