OLH-1456: Add HMRC card to the service dashboard
Add content for the HMRC service card on the service dashboard. The card
will live under "Other services you've used" as opposed to "Your
accounts".

It has a different design from the other services on the service
dashboard due to the requirement to have more content to explain to
users what the card is about. For the time being I handled this by
adding an "isHMRC" flag which is true if the id of a service matches one
of the HMRC client IDs.

For now we should not add this service to the "Services you can use"
list as they'll be in private beta initially.