OLH-1562: Show One Login Home in activity history
We don't currently show sign ins to One Login Home in the history
because there's no content for the client ID in the translation files.
This is intentional because we don't want to show a service card for
Home.

Update the check in the activity history template to use the same test
ID as the backend is expecting, or against the Home app's client ID.