OLH-1580: Sort activity history newest first
Sort the activity history with the newest entry first.

We can't do this in the DynamoDB query because it only supports sorting
using the sort key, and our sort key has to be the event ID because
timestamp may not be unique. This means the entries come back from
DynamoDB in an effectively random order.

The presenter is called before we do pagination, so sorting here means
the order stays consistent across all the displayed pages.