OLH-1588: Filter activity log to items which have content
Filters the activity log down to items which either have a service card,
appear on the 'services you can use' list or are for the One Login Home
RP.

Filtering happens before we decrypt, sort and paginate so this should
also reduce our use of KMS which is likely the slowest part of rendering
the page.

This is needed because the activity log database includes items from RPs
which don't have content eg. the programme's stub RPs. We need to
exclude these to prevent errors when rendering the page. This also
protects us against a new RP onboarding before we've added content for
their service card and activity log. Now if that happens, the activity
log won't display items for that new RP until we have content.