Webchat asset conditional rendering
Moved the contactWebchatEnabled condition into the head block on the
webchat template. The reason for this change is I noticed that a
request was being made for the stylesheet in production and other
environments where the webchat was off.
Moving this conditional block inside the head block instead of
attempting to make the head block conditional on contactWebchatEnabled
seemed to fix the issue. It might have been a case of nunjucks being
quirky.