OLH-1670: Remove 'Find a grant' from services list
The 'Find and apply for a grant' service was wrongly in both the list of
account services and standalone services. This meant that if a user had
the service card they'd see it under both 'Your accounts' and 'Other
services you've used'.

We missed this duplication when we changed these lists in #1192. I've
checked all the other IDs in the lists and there's no other duplicate
entries so this is the only service we need to check.