OLH-1670: Test for no overlap between services lists
We introduced a bug where one service was appearing in both the 'Your
accounts' list and 'Other services you've used' list. This was because
that service's client ID was mistakenly added to both lists in our app
config.

Add a test that checks both lists don't contain any IDs that are in the
other list so we can't make this mistake again.