OLH-1241: Update page title on contact triage page
The title of this page is currently being re-worded in google search
results: from "Contact GOV.UK One Login - GOV.UK One Login" to "Contact
GOV.UK One Login - Sign in".

This might be happening because Google thinks that the title is "keyword
stuffed" due to the repetition of the OL product name. Our content team
have suggested removing the trailing "- GOV.UK One Login" on this page
to help mitigate the problem.
The product name should remain for all other page titles. It should only
be removed on the contact triage page for now.