OLH-1616: Increase frequency of application healthchecks
Reduce the interval between healthcheck requests from our load
balancer to every 15 seconds.

Currently the healthchecks are every 30 seconds and we wait for 2
consecutive failed checks before we mark an instance as unhealthy. This
means that if a container does crash we're waiting for 60 seconds before
starting a new one.

This may be one of the reasons why we're seeing so many failed requests
under load in the performance tests. Making the requests more frequently
means we'll detect crashed containers sooner.