OLH-1709: Add helper function for TxMA header
Add a helper function to get the value of the TxMA device information
header. We'll need to use this values everywhere we send audit events so
we want to have one place we deal with fetching it.

I considered extending the set-local-vars middleware to put this value
in `res.locals` instead, but it's possible that in the future we'll send
audit events in a context that doesn't have clean access to the
`Response` object. I thought it'd be a nicer pattern to call this
function in the controller and pass the result down to the service
that's making the API call or sending the audit event.

There's also a tiny amount of logic here, so a separate function is
easier to unit test.