OLH-1709: Log a warning if the header isn't present
There may be times when this header isn't present - eg. if something
goes wrong with the Cloudfront function that sets it. In those cases we
need to be able to identify from our logs that the problem is happening
and when it started.

If this does become a frequent problem we could also set up a Cloudwatch
metric that counts this log message and use that to trigger an alarm,
but that's out of scope for now.