OLH-1709: Don't return a value if the header is missing
If we don't have a value for the header, we shouldn't include the field
in the audit events at all. We'll deal with that logic at the
controllers where we send the audit events, but we can be clearer that
the value is missing by returning undefined from this function and
leaning on the type system to force us to handle the case where it's
missing.