OLH-1724 Delete account and backchannel logout journeys don't delete user's sessions
The destroyUserSessions function in the session-store module attempts to delete all entries in the user sessions table. However, it did not destroy the Express session which resulted in the session persisting after the user thought they were logged out. Adding a final call to delete the Express session ensured everything was removed. The session-store destroyUserSessions function was refactored to highlight the two distinct actions that it takes, firstly to find all the users records in the session store and delete them and secondly to destroy the Express session.