OLH-1713 Send TxMA header in all calls to Auth's APIs (#1309)
Update the getRequestConfig function to use named parameters rather than positional. This makes it easier to manager, since there are a lot of parameters. We don't need to use null anymore

Use the getTxmaHeader utility function to populate the field, and pass on the txma header

Update tests to check the above

Update types to support the above.