OLH-1748: Fix bug on Change email page
There is an issue on the "Enter your new email address" page where when
there is an error (e.g. the form is submitted with an invalid address),
there is an additional "@" character prepended on the input field.

This bug does not appear to be as a result of an issue on our side, but
rather it seems like it's coming from  the express-validator middleware.
This made it difficult to fix, however I have found a workaround which
seems to eliminate the bug.