OLH-1750: Enter password error state content issue
Fix bug where some of the content is missing on the "Enter password"
page when there is an empty field validation error.
What is happening is that the logic that renders different pieces of
content for different "states" of this page was moved into the
controller, with the strings being passed into the view as local vars.
This works fine in the normal/default state of the view, but it doesn't
work when the form is POSTed and the validateEnterPasswordRequest()
chain function fails with an error, triggering renderBadRequest() which
renders the view again minus those local vars.
My initial intention was to modify renderBadRequest to allow it to
receive locals but I ran into some difficulties with that when
re-rendering in the context of middleware functions. Having spent way
too much time attempting that approach, I switched gears to the
simpler workaround of moving the translation strings into the Nunjucks
template instead.